Refactor, restructured code
