+++ lib/src/utils/utils.dart
@@ -1,0 +1,2 @@
+/// Function to fill with a zero and always show two digits.
+String fill(int n) => n.toString().padLeft(2, "0");
